+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `hooks` subcommand to `smee` command
+- Added `list` subcommand to `hooks` subcommand (#9)
 
 ### Changed